# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.CreateCategoryRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  controller.SwaggerCategoryListData:
    properties:
      list:
        items:
          $ref: '#/definitions/controller.SwaggerCategoryResponse'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  controller.SwaggerCategoryResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: 技术相关文章
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 技术
        type: string
      post_count:
        example: 10
        type: integer
      slug:
        example: ji-shu
        type: string
      sort_order:
        example: 1
        type: integer
    type: object
...
  controller.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  utils.Response:
    properties:
      code:
        description: 状态码，0表示成功
        example: 0
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: your-email@example.com
    name: API Support
  description: 个人博客系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
  /admin/categories:
    post:
      consumes:
      - application/json
      description: 创建新分类（需要管理员权限）
      parameters:
      - description: 分类信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCategoryResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: 分类已存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建分类
      tags:
      - 分类管理
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: 删除分类（需要管理员权限，只能删除没有文章的空白分类）
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 分类不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: 分类下还有文章，无法删除
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除分类
      tags:
      - 分类管理
    put:
      consumes:
      - application/json
      description: 更新分类信息（需要管理员权限）
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCategoryResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 分类不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: 分类已存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新分类
      tags:
      - 分类管理
...
  /categories:
    get:
      consumes:
      - application/json
      description: 分页获取分类列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCategoryListData'
              type: object
      summary: 获取分类列表
      tags:
      - 分类管理
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取分类详情
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCategoryResponse'
              type: object
        "404":
          description: 分类不存在
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取分类详情
      tags:
      - 分类管理
...
  /categories/slug/{slug}:
    get:
      consumes:
      - application/json
      description: 根据Slug获取分类详情
      parameters:
      - description: 分类Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCategoryResponse'
              type: object
        "404":
          description: 分类不存在
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 通过Slug获取分类详情
      tags:
      - 分类管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"